Finstapy now has the methods to like posts, comment on posts, and follow an account.
The AccountPage now has the ability to follow a page, and return how
many followers it has.

The PostElement was given more abilities to comment and like posts, as
well as return an AccountPage class to to the account that posted the
post.

There is now a FollowButtonElement, which has a reference to the follow
button on the AccountPage.