Element was given the SaveInfoButton for the intermediate page between the LoginPage and the HomePage. The alert about turning on notifications was also made into an element to close out of it.
The Locators have been updated to grab the new elements stated in the
Element updates.

Page has been updated to interact with the intermediate page discussed
before. Sleep timers were also added to allow the number of posts loaded
on the screen and the number of buttons that get loaded just a few
seconds later to be closer. A loop was also added to assure that there
is an equal number.